SeldonDeployment/requirements.txt
joblib
pandas
numpy
seldon_core
google-cloud-storage
scikit-learn
nltk